Introduce reusable useOrgQuery composable for dynamic org querying
